feat: orchestrat services
**Registration orchestrator**:
The application state relies on a user having a default character so now
the rpg service will create a character on auth creation

**Task orchestrator**:
Since tasks weret the primary source of events I was having to use an
events service to push notification to the client, which felt like a
very hack workaround. Instead the task orchestrator will receive the
payload necessary for the task and rpg services to give a unified
response to the frontend.