feat: set up task/rpg orchestration
**Task orchestration**
Task resolution (done/undo) now goes through a task orchestrator
The implementation is a little shaky, but it works. The tasks service
validates the task and a non StatusOK response indicates that the RPG
service should revert and changes. This isn't entirely fault free and an
exploration of idempotent, reversible event management will have to be
explored. Nonetheless, quite happy with this change.

**Registration orchestration**
Since users require their character for the app to be in a working
state, they are now assigned a base class upon registration which is
coordinated between the auth service and the rpg service. As with the
task service handling task validation, the auth service takes care of
auth validation. If the auth registration fails, an event is emitted to
delete the userId that was assigned as part of orchestration. This is
delegating some business logic to the orchestrator, but it's another
solution that worksTM

**Refactor**
Services are becoming pretty sizable and are split into app-logic
and user-facing CRUD for personal task and rpg management.

**Internal authorization**
In lieu of setting up a secure interface for inter-service
communication, I've resorted to setting an X-Internal-API header on all
incoming request as false, which is then set to true for services that
need to access internal services. It's not the best way of doing things,
but it works (for now)